requirements.txt
python==3.6.6
Flask==1.0.2
requests==2.18.4 
torch==0.4.1
torchvision==0.2.
pillow==5.2.0
imageio==2.3.0